Real Data Cleaning on Bitcoin and Ether CSV File
This project demonstrates the process of cleaning and analyzing real-world cryptocurrency data (Bitcoin and Ether) from a CSV file. It includes handling missing values, removing outliers, and visualizing the data using Python libraries like Pandas and Matplotlib. The cleaned data is then used to compare mean and median values, as well as to calculate the Interquartile Range (IQR) for dispersion analysis